Refactor transfers
